src/test/java/GameServiceTest.java

import com.magicarena.model.Player;
import com.magicarena.service.GameService;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class GameServiceTest {

    private GameService gameService;
    private Player attacker;
    private Player defender;

    @Before
    public void setUp() {
        gameService = new GameService();
        attacker = new Player(100, 10, 5, "Attacker");
        defender = new Player(100, 10, 5, "Defender");
    }

    @Test
    public void testDamageCalculation() {
        // This test cannot mock dice rolls, but you can set specific known values
        // to the players to ensure certain game mechanics work as expected.
        // You would manually test edge cases like min and max possible rolls.
        int initialHealth = defender.getHealth();
        gameService.performAttack(attacker, defender);

        // Since the actual rolls are unknown, we check conditions:
        assertTrue("Defender's health should be less or equal after an attack",
                defender.getHealth() <= initialHealth);
        assertFalse("Defender should be alive if max damage isn't lethal",
                defender.getHealth() <= 0);
    }
}
